// src/composables/games/was/enums/sound.ts
export const enum WAS_SOUND_ID {
  BGM_MAP = "BGM_MAP",
  BGM_SATAN_CASTLE = "BGM_SATAN_CASTLE",
  BGM_CAVE = "BGM_CAVE",
  BGM_SEA = "BGM_SEA",
  BGM_VILLAGE = "BGM_VILLAGE",
  BGM_MOUNTAIN = "BGM_MOUNTAIN",
  BGM_KINGDOM_CASTLE = "BGM_KINGDOM_CASTLE",
  BGM_BATTLE_1 = "BGM_BATTLE_1",
  BGM_BATTLE_2 = "BGM_BATTLE_2",

  SE_HEAL_1 = "SE_HEAL_1",
  SE_HEAL_2 = "SE_HEAL_2",
  SE_BLOW_2 = "SE_BLOW_2",
  SE_BLOW_3 = "SE_BLOW_3",
  SE_BLOW_4 = "SE_BLOW_4",
  SE_SLASH_1 = "SE_SLASH_1",
  SE_FIRE_1 = "SE_FIRE_1",
  SE_WATER_1 = "SE_WATER_1",
  SE_WIND_1 = "SE_WIND_1",
  SE_THUNDER_1 = "SE_THUNDER_1",
  SE_DARK_1 = "SE_DARK_1",
  SE_SPACIAL_1 = "SE_SPACIAL_1",
}